pyproject.toml
[tool.poetry]
name = "trabalho-final-prog-web"
version = "1.0.0"
description = "Trabalho final da disciplina de Programação para web."
authors = ["ph-cardoso <phlc.cardoso@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlmodel = "^0.0.12"
mysql-connector-python = "^8.2.0"
python-decouple = "^3.8"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
bandit = "^1.7.5"
flake8 = "^6.1.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
